refactor duty cycle and sdr
